Refactor Java UDF JNI modules and harden type coercion
