+++ src/main/java/org/duckdb/UdfOutputAppender.java
@@ -47,9 +47,11 @@ import static org.duckdb.DuckDBBindings.duckdb_vector_assign_string_element_len;
 import static org.duckdb.DuckDBBindings.duckdb_vector_get_column_type;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.nio.LongBuffer;
 import java.sql.SQLException;
+import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
@@ -70,6 +72,13 @@ public final class UdfOutputAppender implements AutoCloseable {
     private static final long UNSIGNED_INT_MAX = 0xFFFF_FFFFL;
     private static final LocalDateTime EPOCH_DATE_TIME = LocalDateTime.ofEpochSecond(0, 0, UTC);
     private static final int MAX_TZ_SECONDS = 16 * 60 * 60 - 1;
+    private static final BigInteger U64_MIN = BigInteger.ZERO;
+    private static final BigInteger U64_MAX = BigInteger.ONE.shiftLeft(64).subtract(BigInteger.ONE);
+    private static final BigInteger U64_MODULUS = BigInteger.ONE.shiftLeft(64);
+    private static final BigInteger HUGEINT_MIN = BigInteger.ONE.shiftLeft(127).negate();
+    private static final BigInteger HUGEINT_MAX = BigInteger.ONE.shiftLeft(127).subtract(BigInteger.ONE);
+    private static final BigInteger UHUGEINT_MIN = BigInteger.ZERO;
+    private static final BigInteger UHUGEINT_MAX = BigInteger.ONE.shiftLeft(128).subtract(BigInteger.ONE);
 
     private final int rowCapacity;
     private final List<Column> columns;
@@ -459,7 +468,7 @@ public final class UdfOutputAppender implements AutoCloseable {
             return;
         case DUCKDB_TYPE_HUGEINT:
         case DUCKDB_TYPE_UHUGEINT:
-            putFixedWidthBytes(col, vectorIdx, requireBytes(value), 16, col.colType.toString());
+            putFixedWidthBytes(col, vectorIdx, requireInt128Bytes(value, col.colType), 16, col.colType.toString());
             return;
         case DUCKDB_TYPE_UUID:
             if (value instanceof UUID) {
@@ -738,9 +747,30 @@ public final class UdfOutputAppender implements AutoCloseable {
             return (BigDecimal) value;
         }
         if (value instanceof Number) {
-            return BigDecimal.valueOf(((Number) value).doubleValue());
+            return toBigDecimal((Number) value);
         }
         throw new IllegalArgumentException("Expected BigDecimal/Number value but got " + value.getClass().getName());
+    }
+
+    private static BigDecimal toBigDecimal(Number value) {
+        if (value instanceof BigInteger) {
+            return new BigDecimal((BigInteger) value);
+        }
+        if (value instanceof Byte || value instanceof Short || value instanceof Integer || value instanceof Long) {
+            return BigDecimal.valueOf(value.longValue());
+        }
+        if (value instanceof Float || value instanceof Double) {
+            try {
+                return BigDecimal.valueOf(value.doubleValue());
+            } catch (NumberFormatException e) {
+                throw new IllegalArgumentException("Expected finite floating-point value for DECIMAL conversion", e);
+            }
+        }
+        try {
+            return new BigDecimal(value.toString());
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("Failed to coerce numeric value to BigDecimal", e);
+        }
     }
 
     private static int requireDateEpochDays(Object value) {
@@ -751,12 +781,23 @@ public final class UdfOutputAppender implements AutoCloseable {
             }
             return (int) days;
         }
+        if (value instanceof java.util.Date) {
+            LocalDate localDate = Instant.ofEpochMilli(((java.util.Date) value).getTime()).atOffset(UTC).toLocalDate();
+            long days = localDate.toEpochDay();
+            if (days < Integer.MIN_VALUE || days > Integer.MAX_VALUE) {
+                throw new IllegalArgumentException("Expected Date epoch day to fit int32 but got " + days);
+            }
+            return (int) days;
+        }
         return (int) requireSignedLongInRange(value, Integer.MIN_VALUE, Integer.MAX_VALUE, "DATE");
     }
 
     private static long requireLongOrTemporal(Object value, DuckDBBindings.CAPIType colType) {
         if (value instanceof Number) {
-            return ((Number) value).longValue();
+            if (colType == DUCKDB_TYPE_UBIGINT) {
+                return requireUnsignedLongBits(value, "UBIGINT");
+            }
+            return requireIntegralLong(value, colType.toString());
         }
         switch (colType) {
         case DUCKDB_TYPE_TIME:
@@ -834,15 +875,13 @@ public final class UdfOutputAppender implements AutoCloseable {
     }
 
     private static long packTimeTzMicros(long micros, int offsetSeconds) {
+        if (offsetSeconds < -MAX_TZ_SECONDS || offsetSeconds > MAX_TZ_SECONDS) {
+            throw new IllegalArgumentException("TIME WITH TIME ZONE offset out of range: " + offsetSeconds +
+                                               " seconds (allowed range: -" + MAX_TZ_SECONDS + ".." + MAX_TZ_SECONDS +
+                                               ")");
+        }
         long normalizedOffset = MAX_TZ_SECONDS - offsetSeconds;
         return ((micros & 0xFFFFFFFFFFL) << 24) | (normalizedOffset & 0xFFFFFFL);
-    }
-
-    private static long requireLong(Object value) {
-        if (!(value instanceof Number)) {
-            throw new IllegalArgumentException("Expected numeric value but got " + value.getClass().getName());
-        }
-        return ((Number) value).longValue();
     }
 
     private static double requireDouble(Object value) {
@@ -853,11 +892,70 @@ public final class UdfOutputAppender implements AutoCloseable {
     }
 
     private static long requireSignedLongInRange(Object value, long min, long max, String typeName) {
-        long num = requireLong(value);
+        long num = requireIntegralLong(value, typeName);
         if (num < min || num > max) {
             throw new IllegalArgumentException("Value out of range for " + typeName + ": " + num);
         }
         return num;
+    }
+
+    private static long requireIntegralLong(Object value, String typeName) {
+        BigInteger integerValue = requireIntegralBigInteger(value, typeName);
+        try {
+            return integerValue.longValueExact();
+        } catch (ArithmeticException e) {
+            throw new IllegalArgumentException("Value out of range for " + typeName + ": " + integerValue, e);
+        }
+    }
+
+    private static long requireUnsignedLongBits(Object value, String typeName) {
+        BigInteger integerValue = requireIntegralBigInteger(value, typeName);
+        if (integerValue.compareTo(U64_MIN) < 0 || integerValue.compareTo(U64_MAX) > 0) {
+            throw new IllegalArgumentException("Value out of range for " + typeName + ": " + integerValue);
+        }
+        if (integerValue.signum() >= 0 && integerValue.bitLength() <= 63) {
+            return integerValue.longValue();
+        }
+        return integerValue.subtract(U64_MODULUS).longValue();
+    }
+
+    private static BigInteger requireIntegralBigInteger(Object value, String typeName) {
+        if (!(value instanceof Number)) {
+            throw new IllegalArgumentException("Expected numeric value for " + typeName + " but got " +
+                                               value.getClass().getName());
+        }
+
+        Number number = (Number) value;
+        if (number instanceof BigInteger) {
+            return (BigInteger) number;
+        }
+        if (number instanceof BigDecimal) {
+            try {
+                return ((BigDecimal) number).toBigIntegerExact();
+            } catch (ArithmeticException e) {
+                throw new IllegalArgumentException("Expected integral value for " + typeName + ": " + number, e);
+            }
+        }
+        if (number instanceof Byte || number instanceof Short || number instanceof Integer || number instanceof Long) {
+            return BigInteger.valueOf(number.longValue());
+        }
+        if (number instanceof Float || number instanceof Double) {
+            double d = number.doubleValue();
+            if (!Double.isFinite(d)) {
+                throw new IllegalArgumentException("Expected finite value for " + typeName + " but got " + d);
+            }
+            try {
+                return BigDecimal.valueOf(d).toBigIntegerExact();
+            } catch (ArithmeticException e) {
+                throw new IllegalArgumentException("Expected integral value for " + typeName + ": " + number, e);
+            }
+        }
+
+        try {
+            return new BigDecimal(number.toString()).toBigIntegerExact();
+        } catch (NumberFormatException | ArithmeticException e) {
+            throw new IllegalArgumentException("Expected integral value for " + typeName + ": " + number, e);
+        }
     }
 
     private static String requireString(Object value) {
@@ -872,6 +970,41 @@ public final class UdfOutputAppender implements AutoCloseable {
             throw new IllegalArgumentException("Expected byte[] value but got " + value.getClass().getName());
         }
         return (byte[]) value;
+    }
+
+    private static byte[] requireInt128Bytes(Object value, DuckDBBindings.CAPIType type) {
+        if (value instanceof byte[]) {
+            return (byte[]) value;
+        }
+        if (value instanceof BigInteger) {
+            return toInt128Bytes((BigInteger) value, type);
+        }
+        throw new IllegalArgumentException("Expected BigInteger/byte[] value for " + type + " but got " +
+                                           value.getClass().getName());
+    }
+
+    private static byte[] toInt128Bytes(BigInteger value, DuckDBBindings.CAPIType type) {
+        if (value == null) {
+            throw new IllegalArgumentException("BigInteger value must not be null");
+        }
+        if (type == DUCKDB_TYPE_HUGEINT) {
+            if (value.compareTo(HUGEINT_MIN) < 0 || value.compareTo(HUGEINT_MAX) > 0) {
+                throw new IllegalArgumentException("Value out of range for HUGEINT: " + value);
+            }
+        } else if (type == DUCKDB_TYPE_UHUGEINT) {
+            if (value.compareTo(UHUGEINT_MIN) < 0 || value.compareTo(UHUGEINT_MAX) > 0) {
+                throw new IllegalArgumentException("Value out of range for UHUGEINT: " + value);
+            }
+        } else {
+            throw new IllegalArgumentException("Int128 conversion is only supported for HUGEINT/UHUGEINT");
+        }
+
+        long lower = value.longValue();
+        long upper = value.shiftRight(64).longValue();
+        ByteBuffer buffer = ByteBuffer.allocate(16).order(java.nio.ByteOrder.nativeOrder());
+        buffer.putLong(lower);
+        buffer.putLong(upper);
+        return buffer.array();
     }
 
     private static void putByte(Column col, long vectorIdx, byte value) {
